Fixed problem in foundation board, details see description
Probelm: not reseting mode back to null when return back to home, and not reseting question model and statictics and mode when finishing a math game. Causing the user being not able to return back to home directly without getting any confirmation dialog at the start page of practise mode and game modes.